SQLiteEmployeeUseCase finished

+++ src/main/java/br/edu/ifsp/addthenewsoul/application/database/SQLiteEmployeeDAO.java
@@ -1,9 +1,8 @@
 package br.edu.ifsp.addthenewsoul.application.database;
 
 import br.edu.ifsp.addthenewsoul.domain.entities.employee.Employee;
+import br.edu.ifsp.addthenewsoul.domain.entities.employee.Role;
 import br.edu.ifsp.addthenewsoul.domain.usecases.employee.EmployeeDAO;
-import javafx.scene.chart.PieChart;
-
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -14,15 +13,37 @@ import java.util.Optional;
 public class SQLiteEmployeeDAO implements EmployeeDAO {
 
     @Override
-    public Optional<Employee> findByRegistrationNumber(Integer registrationNumber) {
+    public Optional<Employee> findByRegistrationNumber(String registrationNumber) {
         String sql = "SELECT * FROM Employee where registrationNumber = ?";
         Employee employee = null;
 
+        try(PreparedStatement stmt = Database.createPreparedStatement(sql)){
+            stmt.setString(1, registrationNumber);
+            ResultSet rs = stmt.executeQuery();
+            if(rs.next()){
+                employee = resultSetToEntity(rs);
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return Optional.ofNullable(employee);
 
     }
 
+    private Employee resultSetToEntity(ResultSet rs) throws SQLException {
+        return new Employee(
+                rs.getString("name"),
+                rs.getString("registrationNumber"),
+                rs.getString("hashPassword"),
+                rs.getString("email"),
+                rs.getString("phone"),
+                Role.toEnum(rs.getString("role"))
+        );
+    }
+
+
     @Override
-    public Integer add(Employee employee) {
+    public String add(Employee employee) {
         String sql = "INSERT INTO Employee(name, registrationNumber, hashPassword, email, phone, role) VALUES (?, ?, ?, ?, ?, ?)";
 
         try(PreparedStatement stmt = Database.createPreparedStatement(sql)) {
@@ -33,8 +54,8 @@ public class SQLiteEmployeeDAO implements EmployeeDAO {
             stmt.setString(5, employee.getPhone());
             stmt.setString(6, employee.getRole().toString());
             stmt.execute();
-            ResultSet rs = stmt.getGeneratedKeys();
-            return rs.getInt(1);
+
+            return employee.getRegistrationNumber();
         } catch (SQLException e) {
             e.printStackTrace();
         }
@@ -47,7 +68,18 @@ public class SQLiteEmployeeDAO implements EmployeeDAO {
         String sql = "SELECT * FROM Employee";
         List<Employee> employees = new ArrayList<>();
 
-        
+        try(PreparedStatement stmt = Database.createPreparedStatement(sql)){
+            ResultSet rs = stmt.executeQuery();
+            while(rs.next()){
+                Employee employee = resultSetToEntity(rs);
+                employees.add(employee);
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return employees;
+
+
     }
 
     @Override
@@ -69,11 +101,21 @@ public class SQLiteEmployeeDAO implements EmployeeDAO {
         return false;
     }
 
+
     @Override
-    public boolean delete(Integer key) {
+    public boolean delete(String registrationNumber) {
+        if (registrationNumber == null)
+            throw new IllegalArgumentException("Registration number must not be null.");
+
         String sql = "DELETE FROM Employee WHERE registrationNumber = ?";
 
-
+        try(PreparedStatement stmt = Database.createPreparedStatement(sql)) {
+            stmt.setString(1, registrationNumber);
+            stmt.execute();
+            return true;
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
         return false;
     }
